优化图片字幕显示

+++ Sources/KSPlayer/Subtitle/AssImageParse.swift
@@ -139,7 +139,7 @@ extension AssImageRenderer: KSSubtitleProtocol {
 /// Pipeline that processed an `ASS_Image` into a ``ProcessedImage`` that can be drawn on the screen.
 public protocol ImagePipelineType {
     init(images: [ASS_Image], boundingRect: CGRect)
-    init(width: Int, height: Int, stride: Int, bitmap: UnsafePointer<UInt8>, palette: UnsafePointer<UInt8>)
+    init(width: Int, height: Int, stride: Int, bitmap: UnsafePointer<UInt8>, palette: UnsafePointer<UInt32>)
     func cgImage(isHDR: Bool, alphaInfo: CGImageAlphaInfo) -> CGImage?
 }
